Add files via upload
For this Udacity project, I was tasked to perform data wrangling (and analyzing and visualizing) of the tweet archive of Twitter user @dog_rates(opens in a new tab), also known as WeRateDogs(opens in a new tab). WeRateDogs is a Twitter account that rates people's dogs with a humorous comment about the dog. This dataset contained over 5000 tweets to perform the wrangling on. 

Project Steps Overview

Your tasks in this project are as follows:

Step 1: Gathering data

Step 2: Assessing data

Step 3: Cleaning data

Step 4: Storing data

Step 5: Analyzing, and visualizing data

Step 6: Reporting

    your data-wrangling efforts
    your data analyses and visualizations